requirements.txt
python3-discogs-client
ascii_magic
colorama
requests